+++ types/config_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/cosmos/cosmos-sdk/crypto/keys/hd"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -43,4 +44,12 @@ func TestSetCoinType(t *testing.T) {
 	require.Equal(t, Bip44CoinType, int(config.GetCoinType()))
 	require.Equal(t, sdk.GetConfig().GetCoinType(), config.GetCoinType())
 	require.Equal(t, sdk.GetConfig().GetFullFundraiserPath(), config.GetFullFundraiserPath())
+}
+
+func TestHDPath(t *testing.T) {
+	params := *hd.NewFundraiserParams(0, Bip44CoinType, 0)
+	// need to prepend "m/" because the below method provided by the sdk does not add the proper prepending
+	hdPath := "m/" + params.String()
+	require.Equal(t, "m/44'/60'/0'/0/0", hdPath)
+	require.Equal(t, hdPath, BIP44HDPath)
 }